Reverting refactor of ServiceOutputParser
