+++ langchain4j/src/main/java/dev/langchain4j/service/DefaultAiServices.java
@@ -15,7 +15,7 @@ import dev.langchain4j.model.output.TokenUsage;
 import dev.langchain4j.rag.AugmentationRequest;
 import dev.langchain4j.rag.AugmentationResult;
 import dev.langchain4j.rag.query.Metadata;
-import dev.langchain4j.service.output.DefaultServiceOutputParser;
+import dev.langchain4j.service.output.ServiceOutputParser;
 import dev.langchain4j.service.output.ServiceOutputParser;
 import dev.langchain4j.service.tool.ToolExecutor;
 
@@ -35,7 +35,7 @@ import static dev.langchain4j.service.TypeUtils.typeHasRawClass;
 
 class DefaultAiServices<T> extends AiServices<T> {
 
-    private final ServiceOutputParser serviceOutputParser = new DefaultServiceOutputParser();
+    private final ServiceOutputParser serviceOutputParser = new ServiceOutputParser();
 
     private static final int MAX_SEQUENTIAL_TOOL_EXECUTIONS = 10;
 
@@ -111,10 +111,9 @@ class DefaultAiServices<T> extends AiServices<T> {
                             userMessage = (UserMessage) augmentationResult.chatMessage();
                         }
 
+                        // TODO give user ability to provide custom OutputParser
                         Type returnType = method.getGenericReturnType();
-                        String outputFormatInstructions = context.hasServiceOutputParser()
-                                ? context.serviceOutputParser.outputFormatInstructions(returnType)
-                                : serviceOutputParser.outputFormatInstructions(returnType);
+                        String outputFormatInstructions = serviceOutputParser.outputFormatInstructions(returnType);
                         String text = userMessage.singleText() + outputFormatInstructions;
                         if (isNotNullOrBlank(userMessage.name())) {
                             userMessage = UserMessage.from(userMessage.name(), text);
@@ -195,9 +194,7 @@ class DefaultAiServices<T> extends AiServices<T> {
 
                         response = Response.from(response.content(), tokenUsageAccumulator, response.finishReason());
 
-                        Object parsedResponse = context.hasServiceOutputParser()
-                                ? context.serviceOutputParser.parse(response, returnType)
-                                : serviceOutputParser.parse(response, returnType);
+                        Object parsedResponse = serviceOutputParser.parse(response, returnType);
                         if (typeHasRawClass(returnType, Result.class)) {
                             return Result.builder()
                                     .content(parsedResponse)
